+++ 03-Prefix-Sum/prefix-sum/src/rangesum/Solution.java
@@ -11,17 +11,17 @@ public class Solution {
         this.b = b;
     }
 
-    public int[] rangeSum() {
-        int[] result = new int[b.length];
+    public long[] rangeSum() {
+        long[] result = new long[b.length];
         for(int i = 0 ; i < b.length; i++) {
-            int sum = sumResult(arr, b[i][0], b[i][1]);
+            long sum = sumResult(arr, b[i][0], b[i][1]);
             result[i] = sum;
         }
         return result;
     }
 
-    private int sumResult(int[] arr, int s, int e) {
-        int sum = 0;
+    private long sumResult(int[] arr, int s, int e) {
+        long sum = 0;
         for(int i = s; i <= e; i++) { sum += arr[i]; }
         return sum;
     }
